+++ jamf_upload_lib/actions.py
@@ -50,3 +50,55 @@ def status_check(r, endpoint_type, obj_name):
         print(
             f"WARNING: {endpoint_type} '{obj_name}' upload failed with status code {r.status_code}"
         )
+
+
+def confirm(prompt=None, default=False):
+    """prompts for yes or no response from the user. Returns True for yes and
+    False for no.
+
+    'default' should be set to the default value assumed by the caller when
+    user simply types ENTER.
+
+    Examples:
+
+    confirm(prompt='Create Directory?', default=True)
+    Create Directory? [y]|n:
+    True
+
+    confirm(prompt='Create Directory?', default=False)
+    Create Directory? [n]|y:
+    False
+
+    confirm(prompt='Create Directory?', default=False)
+    Create Directory? [n]|y: y
+    True
+
+    Source:
+    https://code.activestate.com/recipes/541096-prompt-the-user-for-confirmation/
+    """
+
+    if prompt is None:
+        prompt = "Confirm"
+
+    if default:
+        prompt = "%s [%s]|%s: " % (prompt, "y", "n")
+    else:
+        prompt = "%s [%s]|%s: " % (prompt, "n", "y")
+
+    while True:
+        answer = input(prompt)
+        if not answer:
+            return default
+        if answer not in ["y", "Y", "n", "N"]:
+            print("Please enter Y or N.")
+            continue
+        if answer == "y" or answer == "Y":
+            return True
+        if answer == "n" or answer == "N":
+            return False
+
+
+# def days_between(d1, d2):
+#     d1 = datetime.strptime(d1, "%Y-%m-%d")
+#     d2 = datetime.strptime(d2, "%Y-%m-%d")
+#     return abs((d2 - d1).days)